+++ src/features/goblins/tailor/Tailor.tsx
@@ -1,20 +1,14 @@
-import React, { useContext } from "react";
-import { useActor } from "@xstate/react";
-import { Modal } from "react-bootstrap";
+import React from "react";
 
-import { Context } from "features/game/GameProvider";
-
-import tailor from "assets/buildings/tailor.gif";
+import goblin from "assets/npcs/goblin.gif";
 import flag from "assets/nfts/flags/sunflower_flag.gif";
 
 import { GRID_WIDTH_PX } from "features/game/lib/constants";
 import { Action } from "components/ui/Action";
-import { TailorSale } from "./components/TailorSale";
 import { tailorAudio } from "lib/utils/sfx";
+import { ItemsModal } from "./ItemsModal";
 
 export const Tailor: React.FC = () => {
-  // const { gameService } = useContext(Context);
-  // const [gameState] = useActor(gameService);
   const [isOpen, setIsOpen] = React.useState(false);
 
   const openTailor = () => {
@@ -26,16 +20,16 @@ export const Tailor: React.FC = () => {
     <div
       className="absolute"
       style={{
-        width: `${GRID_WIDTH_PX * 3.5}px`,
-        right: `${GRID_WIDTH_PX * 6}px`,
-        top: `${GRID_WIDTH_PX * 36}px`,
+        width: `${GRID_WIDTH_PX * 0.9}px`,
+        right: `${GRID_WIDTH_PX * 7.9}px`,
+        top: `${GRID_WIDTH_PX * 6.6}px`,
       }}
     >
       <div className="cursor-pointer hover:img-highlight">
-        <img src={tailor} className="w-full" onClick={openTailor} />
+        <img src={goblin} className="w-full" onClick={openTailor} />
         {
           <Action
-            className="absolute -bottom-7 -left-2"
+            className="absolute -bottom-10 -left-8"
             text="Tailor"
             icon={flag}
             onClick={openTailor}
@@ -44,9 +38,7 @@ export const Tailor: React.FC = () => {
       </div>
 
       {isOpen && (
-        <Modal centered show={isOpen} onHide={() => setIsOpen(false)}>
-          <TailorSale onClose={() => setIsOpen(false)} />
-        </Modal>
+        <ItemsModal isOpen={isOpen} onClose={() => setIsOpen(false)} />
       )}
     </div>
   );
